Preview Drawing with GameObject/SpriteRenderer
Converted everything over to work with a GameObject that has a
SpriteRenderer component that exists on a sorting layer above the layer
we're currently working on.

Also changed how the MapEditorWindow interacts with the MapEditor.
Instead of public static methods, it now just changed properties of the
MapEditor, which triggers the appropriate preview object changes.

Todo: Add all the MapEditorWindow interface stuff that lets us change
floors, select prefabs, load/save maps, etc.