PrefabManager and Window Loading and Preview Generation
Added functionallity to the PrefabManager to generate seperate
Texture2D's for each prefab, so that we can display them in the
PrefabManagerWindow, and so we can display the currently selected
prefab's texture in the MapEditorWindow.

Next to do: Clean up the display of prefab buttons, so they display row
by row.