Prefab Icon Rows
Changed how PrefabManager loads. Instead of using FindObjectsOfTypeAll,
we now use LoadAll and a preconfigured array of folders. This solves the
issue of having to look at prefabs in folders first, so that unity
'indexes' them. Now works correctly each time we press the button.

Added functionality to display preview textures in rows, and to
dynamically adjust with the window size and the size of the preview
texture.

TODO: Prevent unity from cleaning up 'leaked textures' whenever we save
the scene.