MapEditor UI (New/Load/Save/SaveAs)
MapEditorWindow: enabled functionality for creating, loading, save, and
save as-ing maps.  Added map statistics.
MapEditor: backend functionality for UI changes, as well as performance
improvement for preview squares. Now repainting the scene when there's a
change.
InstanceManager: added functionality to clear the manager.

Note: There's the problem of the MapEditor's map getting unlinked if we
change the code for the map editor. Additionally, all the map properties
are tied to the MapEditor, which is wrong.

Considering abstracting the map from its data.